Use more universal name for the handler and move them into the configuration package
