+++ src/main/java/com/md/scheduler/configuration/exception/ApiExceptionHandler.java
@@ -1,6 +1,7 @@
-package com.md.scheduler.users.registration;
+package com.md.scheduler.configuration.exception;
 
-import com.md.scheduler.configuration.exception.ApiException;
+import com.md.scheduler.users.registration.UserAlreadyExistAuthenticationException;
+import org.springframework.context.support.DefaultMessageSourceResolvable;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -10,14 +11,14 @@ import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
 @RestControllerAdvice
-class UserAlreadyExistAuthenticationExceptionHandler extends ResponseEntityExceptionHandler {
+class ApiExceptionHandler extends ResponseEntityExceptionHandler {
 
     @ExceptionHandler(UserAlreadyExistAuthenticationException.class)
-    ResponseEntity<Object> handler(UserAlreadyExistAuthenticationException e, WebRequest request) {
+    ResponseEntity<Object> handler(UserAlreadyExistAuthenticationException ex, WebRequest request) {
         return new ResponseEntity<>(
                         ApiException.builder()
                                 .type("/errors/user-already-exist")
-                                .title(e.getMessage())
+                                .title(ex.getMessage())
                                 .status(HttpStatus.CONFLICT.value())
                                 .detail("Select other username, because your choice is already in use in our system")
                                 .instance(request.getDescription(false))
