+++ src/main/java/com/md/scheduler/users/registration/UserAlreadyExistAuthenticationException.java
@@ -2,7 +2,7 @@ package com.md.scheduler.users.registration;
 
 import javax.naming.AuthenticationException;
 
-class UserAlreadyExistAuthenticationException extends AuthenticationException {
+public class UserAlreadyExistAuthenticationException extends AuthenticationException {
 
     UserAlreadyExistAuthenticationException(final String msg) {
         super(msg);
